Refactoring the all modes into its own meta and merge the ResMeta and GroupMeta to one.
