--- internal/meta/terraform_hack.go
@@ -1,44 +1,0 @@
-package meta
-
-import (
-	"fmt"
-)
-
-// lifecycleAddon adds lifecycle meta arguments for some identified resources, which are mandatory to make them usable.
-func (meta Meta) lifecycleAddon(configs ConfigInfos) (ConfigInfos, error) {
-	out := make(ConfigInfos, len(configs))
-	for i, cfg := range configs {
-		switch cfg.TFAddr.Type {
-		case "azurerm_application_insights_web_test":
-			if err := hclBlockAppendLifecycle(cfg.hcl.Body().Blocks()[0].Body(), []string{"tags"}); err != nil {
-				return nil, fmt.Errorf("azurerm_application_insights_web_test: %v", err)
-			}
-		}
-		out[i] = cfg
-	}
-	return out, nil
-}
-
-func (meta Meta) addDependency(configs ConfigInfos) (ConfigInfos, error) {
-	if err := configs.AddDependency(); err != nil {
-		return nil, err
-	}
-
-	var out ConfigInfos
-
-	configSet := map[string]ConfigInfo{}
-	for _, cfg := range configs {
-		configSet[cfg.AzureResourceID.String()] = cfg
-	}
-
-	for _, cfg := range configs {
-		if len(cfg.DependsOn) != 0 {
-			if err := hclBlockAppendDependency(cfg.hcl.Body().Blocks()[0].Body(), cfg.DependsOn, configSet); err != nil {
-				return nil, err
-			}
-		}
-		out = append(out, cfg)
-	}
-
-	return out, nil
-}
