+++ internal/test/query/query_test.go
@@ -83,7 +83,7 @@ resource "azurerm_subnet" "test" {
 
 	// Import in non-recursive mode
 	aztfyDir := t.TempDir()
-	cfg := config.GroupConfig{
+	cfg := config.Config{
 		CommonConfig: config.CommonConfig{
 			SubscriptionId: os.Getenv("ARM_SUBSCRIPTION_ID"),
 			OutputDir:      aztfyDir,
