+++ internal/test/resmap/e2e_cases_test.go
@@ -66,7 +66,7 @@ func runCase(t *testing.T, d test.Data, c cases.Case) {
 	require.NoError(t, err)
 	require.NoError(t, utils.WriteFileSync(mapFile, bMapping, 0644))
 
-	cfg := config.GroupConfig{
+	cfg := config.Config{
 		CommonConfig: config.CommonConfig{
 			SubscriptionId: os.Getenv("ARM_SUBSCRIPTION_ID"),
 			OutputDir:      aztfyDir,
