+++ internal/ui/aztfyclient/client.go
@@ -8,7 +8,7 @@ import (
 	tea "github.com/charmbracelet/bubbletea"
 )
 
-type NewClientMsg meta.GroupMeta
+type NewClientMsg meta.Meta
 
 type ErrMsg error
 
@@ -52,13 +52,13 @@ type CleanTFStateMsg struct {
 	Addr string
 }
 
-func NewClient(meta meta.GroupMeta) tea.Cmd {
+func NewClient(meta meta.Meta) tea.Cmd {
 	return func() tea.Msg {
 		return NewClientMsg(meta)
 	}
 }
 
-func Init(c meta.GroupMeta) tea.Cmd {
+func Init(c meta.Meta) tea.Cmd {
 	return func() tea.Msg {
 		err := c.Init()
 		if err != nil {
@@ -68,7 +68,7 @@ func Init(c meta.GroupMeta) tea.Cmd {
 	}
 }
 
-func ListResource(c meta.GroupMeta) tea.Cmd {
+func ListResource(c meta.Meta) tea.Cmd {
 	return func() tea.Msg {
 		list, err := c.ListResource()
 		if err != nil {
@@ -84,13 +84,13 @@ func ShowImportError(item meta.ImportItem, idx int, l meta.ImportList) tea.Cmd {
 	}
 }
 
-func StartImport(c meta.GroupMeta, l meta.ImportList) tea.Cmd {
+func StartImport(c meta.Meta, l meta.ImportList) tea.Cmd {
 	return func() tea.Msg {
 		return StartImportMsg{List: l}
 	}
 }
 
-func ImportOneItem(c meta.GroupMeta, item meta.ImportItem) tea.Cmd {
+func ImportOneItem(c meta.Meta, item meta.ImportItem) tea.Cmd {
 	return func() tea.Msg {
 		if !item.Skip() && !item.Imported {
 			c.Import(&item)
@@ -108,7 +108,7 @@ func FinishImport(l meta.ImportList) tea.Cmd {
 	}
 }
 
-func GenerateCfg(c meta.GroupMeta, l meta.ImportList) tea.Cmd {
+func GenerateCfg(c meta.Meta, l meta.ImportList) tea.Cmd {
 	return func() tea.Msg {
 		if err := c.GenerateCfg(l); err != nil {
 			return ErrMsg(err)
@@ -117,7 +117,7 @@ func GenerateCfg(c meta.GroupMeta, l meta.ImportList) tea.Cmd {
 	}
 }
 
-func ExportResourceMapping(c meta.GroupMeta, l meta.ImportList) tea.Cmd {
+func ExportResourceMapping(c meta.Meta, l meta.ImportList) tea.Cmd {
 	return func() tea.Msg {
 		if err := c.ExportResourceMapping(l); err != nil {
 			return ErrMsg(err)
@@ -126,7 +126,7 @@ func ExportResourceMapping(c meta.GroupMeta, l meta.ImportList) tea.Cmd {
 	}
 }
 
-func ExportSkippedResources(c meta.GroupMeta, l meta.ImportList) tea.Cmd {
+func ExportSkippedResources(c meta.Meta, l meta.ImportList) tea.Cmd {
 	return func() tea.Msg {
 		if err := c.ExportSkippedResources(l); err != nil {
 			return ErrMsg(err)
